backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
langchain
langchain-openai
langchain-groq
pinecone
pymongo
motor
arxiv
praw
python-dotenv
pytest
httpx
aiohttp